+++ source/samples/OpenGL/OpenGLSample.cpp
@@ -10,7 +10,7 @@
 #include <Gwork/Skins/Simple.h>
 #include <Gwork/Skins/TexturedBase.h>
 #include <Gwork/Test/Test.h>
-#include <Gwork/Input/OpenGL.h>
+#include <Gwork/Input/GLFW3.h>
 #ifdef USE_DEBUG_FONT
 #   include <Gwork/Renderers/OpenGL_DebugFont.h>
 #else
@@ -18,11 +18,11 @@
 #endif
 #include <GLFW/glfw3.h>
 
-static Gwk::Input::GLFW GworkInput;
+static Gwk::Input::GLFW3 GworkInput;
 
 static void key_callback(GLFWwindow* window, int key, int scancode, int action, int mods)
 {
-    const Gwk::Input::GLFW::KeyEvent evt = { key, scancode, action, mods };
+    const Gwk::Input::GLFW3::KeyEvent evt = { key, scancode, action, mods };
     GworkInput.ProcessKeyEvent(evt);
 }